refactor: ManagementGroupWithIdコンポーネントを[id].tsxに移動
